rust: SMP: add test code for SMP
The test function is defined as executing tasks in parallel on multiple
CPU cores. To correctly perform SMP testing, you need to set the kernel
startup parameter `oobcpus_arg` in the `kernel/rros/init.rs` file, and
also use `-smp` in the `qemu` startup parameter to define the corresponding
number of SMP cores.

https://github.com/BUPT-OS/RROS/issues/40

Signed-off-by: JiajunDu <wonderboy512@163.com>